pyproject.toml
[tool.poetry]
name = "django-file-tools"
version = "0.1.0"
description = "Tools for working with files in Django"
authors = ["Enrico Barzetti"]
readme = "README.md"
homepage = "https://github.com/enricobarzetti/django-file-tools"
license = "mit"
packages = [
    { include = "django_file_tools" },
]

[tool.poetry.dependencies]
python = "^3.9"
django = ">=3.0"
django-storages = "^1.14"
djangorestframework = "^3.12.0"
boto3 = "^1.34.12"
django-environ = "^0.11.2"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
